Fixed passing of 2d array to java method
For some reason, jpype does not convert a JArray as expected, you must manually create a list of lists (nested list) in Python and convert all the items to a java Float using the jpype wrappers to the java 'Float' type. Then pass that list to the java method. (The method is expecting type 'float[][]', i.e. a 2D array of floats.)